HW02

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepositoryImpl.java
@@ -3,11 +3,17 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.toList;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -29,18 +35,31 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        if (repository.containsKey(id)&&repository.get(id).getUserId().equals(userId)){
+            repository.remove(id);
+            return true;
+        }
+        //throw new ru.javawebinar.topjava.util.exception.NotFoundException("Чужая еда!");
+        return false;
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, int userId) {
+        if (repository.get(id).getUserId().equals(userId))
         return repository.get(id);
+        throw new ru.javawebinar.topjava.util.exception.NotFoundException("Meal not found.");
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        List<Meal> allMeal=
+                repository.values().stream()
+                //.filter(meal -> meal.getUserId().equals(userId))
+                .sorted(Comparator.comparing(Meal::getDateTime)).collect(Collectors.toList());//.collect(toList());
+        if (allMeal.equals(null))
+            throw new ru.javawebinar.topjava.util.exception.NotFoundException("Meal(s) not found");
+        return allMeal;
     }
 }
